% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elem_detect_outlier.R
\name{elem_detect_outlier}
\alias{elem_detect_outlier}
\title{Detect univariate outlier by boxplot method.}
\usage{
elem_detect_outlier(x, coef = 1.5, na.rm = FALSE, ...)
}
\arguments{
\item{x}{(numerical): Input data. Must be a numerical array, generally containg concentration values, on which should be detected the anomalous values.}

\item{coef}{(numerical): A number, generally 1.5 ("first order anomaly"), 3.0 ("second order anomaly") or 4.5 ("thrid order anomaly"), indicating the coeficient to apply on the multiplication to determine the anomalous values.}

\item{na.rm}{(logical): If TRUE, all NA values would be removed before the calculation.}

\item{...}{(ellipsis): any additional variable;}
}
\value{
factor with 3 levels: background, outlier and missing.
}
\description{
Detect univariate outlier by boxplot method.
}
