% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elem_imput.R
\name{elem_imput}
\alias{elem_imput}
\title{Impute values lower than limit of detection by replacing it for the multiplication of coef*LOD}
\usage{
elem_imput(df, coef = 0.5, ...)
}
\arguments{
\item{df}{(data.frame): a data.frame containing the elements concentrations}

\item{coef}{(double): default = 0.5. A number, generally lower than 1, to be multiplied to the lower detection limit value.}

\item{...}{(ellipsis): any additional variable;}
}
\value{
data.frame
}
\description{
Impute values lower than limit of detection by replacing it for the multiplication of coef*LOD
}
