% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elem_norm.R
\name{elem_norm}
\alias{elem_norm}
\title{Normalize your data into the minmax or centered log ratio distribution}
\usage{
elem_norm(df, method = "minmax", keep = NULL, ...)
}
\arguments{
\item{df}{(data.frame): a data.frame object containing at least one numeric variable to be normalized.}

\item{method}{(character): the method choosen for normalization <default = minmax>. Must be one of the three options: "minmax" (defines a range greater than or equal to 0 and less than or equal to 1), "clr" (centered log ratio transformation) and "log10" (logarithm base 10).}

\item{keep}{(character): a character vector containing the names of numeric columns to be preserved of the normalization.}

\item{...}{(ellipsis): any additional variable;}
}
\value{
data.frame
}
\description{
Normalize your data into the minmax or centered log ratio distribution
}
