% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elem_select.R
\name{elem_select}
\alias{elem_select}
\title{Select elements by the proportion of non-censured values}
\usage{
elem_select(df, cut = 0.75, ...)
}
\arguments{
\item{df}{(data.frame): a 'data.frame' or flat object with numerical variables}

\item{cut}{(double): A number <default = 0.75>, generally lower than 1, to be multiplied to the lower detection limit value.}

\item{...}{(ellipsis): any additional variable;}
}
\value{
data.frame
}
\description{
Select elements by the proportion of non-censured values
}
